requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
google-genai>=1.0.0
google-cloud-bigquery[pandas]>=3.10.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.65.0
pandas>=2.0.0
